+++ client/src/containers/analysis-visualization/analysis-table/component.tsx
@@ -169,7 +169,7 @@ const AnalysisTable: React.FC = () => {
                 title={original.name}
               >
                 {/* @ts-expect-error the parent and the children have different data types */}
-                {`${original.name} ${depth === 0 && `(${original.metadata.unit})`}`}
+                {`${original.name} ${depth === 0 ? `(${original.metadata.unit})` : ''}`}
               </span>
             </>
           );